Clean code (v3)
More modularity (Modular programming).

JavaFX library included inside a folder of this project.

module-info.java ready for exporting the app to JAR.

Change app behaviour when no directory is selected.

HTML, CSS and pictures added to a specific folder: resources.

Using File.separator so the app works in any SO.